HW7: Скорректировал расположение файлов types.ts
