+++ src/components/catalogCard/types.ts
@@ -1,0 +1,5 @@
+import { ICatalogItem } from "@slices/catalogSlice/types";
+
+export interface ICatalogItemProps {
+    item: ICatalogItem;
+  }
