+++ src/slices/errorSlice/types.ts
@@ -1,0 +1,9 @@
+export interface IError {
+    message: string;
+    code?: string;
+}
+
+export interface IErrorState {
+    errorQueue: Error[];
+    currentError: Error | null;
+}
