+++ src/slices/productSlice/types.ts
